Refactored the audioplayer to its own file and added labels to the answer radiobuttons.
